Refactor WinRT method and property projection classes
