+++ tool/generator/lib/generator.dart
@@ -25,12 +25,13 @@ export 'src/projection/struct.dart';
 export 'src/projection/test_interface.dart';
 export 'src/projection/type.dart';
 export 'src/projection/utils.dart';
-export 'src/projection/winrt_class.dart';
-export 'src/projection/winrt_get_property.dart';
-export 'src/projection/winrt_interface.dart';
-export 'src/projection/winrt_method.dart';
-export 'src/projection/winrt_property.dart';
-export 'src/projection/winrt_set_property.dart';
+export 'src/projection/winrt/winrt_class.dart';
+export 'src/projection/winrt/winrt_get_property.dart';
+export 'src/projection/winrt/winrt_interface.dart';
+export 'src/projection/winrt/winrt_method.dart';
+export 'src/projection/winrt/winrt_parameter.dart';
+export 'src/projection/winrt/winrt_property.dart';
+export 'src/projection/winrt/winrt_set_property.dart';
 
 export 'src/shared/exclusions.dart';
 export 'src/shared/false_properties.dart';
