+++ tool/generator/lib/src/projection/winrt/winrt_class.dart
@@ -1,6 +1,6 @@
 import 'package:winmd/winmd.dart';
 
-import 'utils.dart';
+import '../utils.dart';
 import 'winrt_interface.dart';
 
 class WinRTClassProjection extends WinRTInterfaceProjection {
